Added loadOrders and DataAccess class
NOTE! I think this will work, as I'm updating the .classpath file as
well, but I'm adding an external library, so you may have to add the
sqlite-jdbc-3.8.11.1.jar to your class path in Preferences for it to
work. Also the order number will increment every time you open the
program. This is fine, as it re-adds the dummy data every time and in
the future will will want every order to have its own order number, even
if previous orders have been deleted.